requirements.txt
python-telegram-bot==20.3
requests==2.31.0
apscheduler==3.10.1
python-dotenv==1.0.0
